mini project/anemia-screening/requirements.txt
Flask==2.3.3
opencv-python==4.8.1.78
Pillow==10.0.1
